Remove COSE Algorithms for "[ALG] with key derived by [ARKG]"
These shouldn't be needed since the "derived by ARKG" part is already expressed
in the key type. The signature verifier does not need to do anything special to
verify these signatures, as it is assumed the verifier receives the derived
public key already blinded and can use it to verify the signature as normal.